refactor: split fetch package into multiple files
Split out the fetch package into multiple files,
and extracted Getter interface into its own package
